test: add read-only example tests
Also refactored getRoleBinding function out into its own
helpers_test file.
